доработки с типами. 1. domain.d.ts => domain.ts 2. скрипт для генерации типов в package.json 3. перевел пример на ts 4. экспорт типов
