+++ example/index.ts
@@ -1,15 +1,16 @@
-const OpenAPI = require('@tinkoff/invest-openapi-js-sdk');
-
+import OpenAPI, { MarketInstrument } from '@tinkoff/invest-openapi-js-sdk';
 const apiURL = 'https://api-invest.tinkoff.ru/openapi';
 const sandboxApiURL = 'https://api-invest.tinkoff.ru/openapi/sandbox/';
 const socketURL = 'wss://api-invest.tinkoff.ru/openapi/md/v1/md-openapi/ws';
 const secretToken = process.env.TOKEN; // токен для боевого api
 const sandboxToken = process.env.SANDBOX_TOKEN; // токен для сандбокса
-const api = new OpenAPI({ apiURL: sandboxApiURL, secretToken: sandboxToken, socketURL });
+const api = new OpenAPI({ apiURL: sandboxApiURL, secretToken: sandboxToken as string, socketURL });
 
 !(async function run() {
   await api.sandboxClear();
-  const { figi } = await api.searchOne({ ticker: 'AAPL' });
+
+  const marketInstrument = await api.searchOne({ ticker: 'AAPL' }) as MarketInstrument;
+  const { figi } = marketInstrument;
   console.log(await api.setCurrenciesBalance({ currency: 'USD', balance: 1000 })); // 1000$ на счет
   console.log(await api.portfolioCurrencies());
   console.log(await api.instrumentPortfolio({ figi })); // В портфеле ничего нет
