+++ example/package.json
@@ -4,11 +4,15 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "ts-node index.ts"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@tinkoff/invest-openapi-js-sdk": "^1.2.3"
+    "@tinkoff/invest-openapi-js-sdk": "^1.2.4"
+  },
+  "devDependencies": {
+    "ts-node": "^8.8.1",
+    "typescript": "^3.8.3"
   }
 }
